requirements.txt
python_hcl2==5.1.1
setuptools==75.6.0
